+++ src/DurableTask.SqlServer/SqlTypes/OrchestrationEventSqlType.cs
@@ -30,6 +30,7 @@ namespace DurableTask.SqlServer.SqlTypes
             new SqlMetaData("PayloadID", SqlDbType.UniqueIdentifier),
             new SqlMetaData("ParentInstanceID", SqlDbType.VarChar, 100),
             new SqlMetaData("Version", SqlDbType.VarChar, 100),
+            new SqlMetaData("TraceContext", SqlDbType.VarChar, 800),
         };
 
         static class ColumnOrdinals
@@ -48,6 +49,7 @@ namespace DurableTask.SqlServer.SqlTypes
             public const int PayloadId = 10;
             public const int ParentInstanceID = 11;
             public const int Version = 12;
+            public const int TraceContext = 13;
         }
 
         public static SqlParameter AddOrchestrationEventsParameter(
@@ -149,6 +151,7 @@ namespace DurableTask.SqlServer.SqlTypes
 
             record.SetSqlString(ColumnOrdinals.ParentInstanceID, SqlUtils.GetParentInstanceId(msg.Event));
             record.SetSqlString(ColumnOrdinals.Version, SqlUtils.GetVersion(msg.Event));
+            record.SetSqlString(ColumnOrdinals.TraceContext, SqlUtils.GetTraceContext(msg.Event));
 
             return record;
         }