+++ src/DurableTask.SqlServer/SqlTypes/TaskEventSqlType.cs
@@ -29,6 +29,7 @@ namespace DurableTask.SqlServer.SqlTypes
             new SqlMetaData("PayloadText", SqlDbType.VarChar, -1 /* max */),
             new SqlMetaData("PayloadID", SqlDbType.UniqueIdentifier),
             new SqlMetaData("Version", SqlDbType.VarChar, 100),
+            new SqlMetaData("TraceContext", SqlDbType.VarChar, 800),
         };
 
         static class ColumnOrdinals
@@ -46,6 +47,7 @@ namespace DurableTask.SqlServer.SqlTypes
             public const int PayloadText = 9;
             public const int PayloadId = 10;
             public const int Version = 11;
+            public const int TraceContext = 12;
         }
 
         public static SqlParameter AddTaskEventsParameter(
@@ -137,7 +139,7 @@ namespace DurableTask.SqlServer.SqlTypes
             // to pre-lock task work items for this particular node.
 
             record.SetSqlString(ColumnOrdinals.Version, SqlUtils.GetVersion(msg.Event));
-
+            record.SetSqlString(ColumnOrdinals.TraceContext, SqlUtils.GetTraceContext(msg.Event));
             return record;
         }
     }