+++ src/DurableTask.SqlServer/SqlUtils.cs
@@ -10,10 +10,12 @@ namespace DurableTask.SqlServer
     using System.Data.SqlTypes;
     using System.Diagnostics;
     using System.Linq;
+    using System.Text;
     using System.Threading;
     using System.Threading.Tasks;
     using DurableTask.Core;
     using DurableTask.Core.History;
+    using DurableTask.Core.Tracing;
     using Microsoft.Data.SqlClient;
     using Microsoft.Data.SqlClient.Server;
     using SemVersion;
@@ -21,12 +23,13 @@ namespace DurableTask.SqlServer
     static class SqlUtils
     {
         static readonly Random random = new Random();
+        static readonly char[] TraceContextSeparators = new char[] { '\n' };
 
         public static string? GetStringOrNull(this DbDataReader reader, int columnIndex)
         {
             return reader.IsDBNull(columnIndex) ? null : reader.GetString(columnIndex);
         }
-        
+
         public static TaskMessage GetTaskMessage(this DbDataReader reader)
         {
             return new TaskMessage
@@ -61,6 +64,7 @@ namespace DurableTask.SqlServer
                     historyEvent = new EventRaisedEvent(eventId, GetPayloadText(reader))
                     {
                         Name = GetName(reader),
+                        ParentTraceContext = GetTraceContext(reader),
                     };
                     break;
                 case EventType.EventSent:
@@ -78,10 +82,18 @@ namespace DurableTask.SqlServer
                         orchestrationStatus: OrchestrationStatus.Completed);
                     break;
                 case EventType.ExecutionFailed:
+                    string? executionFailedResult = null;
+                    if (!TryGetFailureDetails(reader, out FailureDetails? executionFailedDetails))
+                    {
+                        // Fall back to the old behavior
+                        executionFailedResult = GetPayloadText(reader);
+                    }
+
                     historyEvent = new ExecutionCompletedEvent(
                         eventId,
-                        result: GetPayloadText(reader),
-                        orchestrationStatus: OrchestrationStatus.Failed);
+                        result: executionFailedResult,
+                        orchestrationStatus: OrchestrationStatus.Failed,
+                        failureDetails: executionFailedDetails);
                     break;
                 case EventType.ExecutionStarted:
                     historyEvent = new ExecutionStartedEvent(eventId, GetPayloadText(reader))
@@ -94,6 +106,7 @@ namespace DurableTask.SqlServer
                         },
                         Tags = null, // TODO
                         Version = GetVersion(reader),
+                        ParentTraceContext = GetTraceContext(reader),
                     };
                     string? parentInstanceId = GetParentInstanceId(reader);
                     if (parentInstanceId != null)
@@ -121,7 +134,7 @@ namespace DurableTask.SqlServer
                     historyEvent = new OrchestratorStartedEvent(eventId);
                     break;
                 case EventType.SubOrchestrationInstanceCompleted:
-                    historyEvent = new SubOrchestrationInstanceCompletedEvent(eventId, GetTaskId(reader), GetPayloadText(reader));
+                    historyEvent = new SubOrchestrationInstanceCompletedEvent(eventId: -1, GetTaskId(reader), GetPayloadText(reader));
                     break;
                 case EventType.SubOrchestrationInstanceCreated:
                     historyEvent = new SubOrchestrationInstanceCreatedEvent(eventId)
@@ -133,24 +146,44 @@ namespace DurableTask.SqlServer
                     };
                     break;
                 case EventType.SubOrchestrationInstanceFailed:
+                    string? subOrchFailedReason = null;
+                    string? subOrchFailedDetails = null;
+                    if (!TryGetFailureDetails(reader, out FailureDetails? subOrchFailureDetails))
+                    {
+                        // Fall back to the old behavior
+                        subOrchFailedReason = GetReason(reader);
+                        subOrchFailedDetails = GetPayloadText(reader);
+                    }
+
                     historyEvent = new SubOrchestrationInstanceFailedEvent(
-                        eventId,
+                        eventId: -1,
                         taskScheduledId: GetTaskId(reader),
-                        reason: GetReason(reader),
-                        details: GetPayloadText(reader));
+                        reason: subOrchFailedReason,
+                        details: subOrchFailedDetails,
+                        failureDetails: subOrchFailureDetails);
                     break;
                 case EventType.TaskCompleted:
                     historyEvent = new TaskCompletedEvent(
-                        eventId,
+                        eventId: -1,
                         taskScheduledId: GetTaskId(reader),
                         result: GetPayloadText(reader));
                     break;
                 case EventType.TaskFailed:
+                    string? taskFailedReason = null;
+                    string? taskFailedDetails = null;
+                    if (!TryGetFailureDetails(reader, out FailureDetails? taskFailureDetails))
+                    {
+                        // Fall back to the old behavior
+                        taskFailedReason = GetReason(reader);
+                        taskFailedDetails = GetPayloadText(reader);
+                    }
+
                     historyEvent = new TaskFailedEvent(
-                        eventId,
+                        eventId: -1,
                         taskScheduledId: GetTaskId(reader),
-                        reason: GetReason(reader),
-                        details: GetPayloadText(reader));
+                        reason: taskFailedReason,
+                        details: taskFailedDetails,
+                        failureDetails: taskFailureDetails);
                     break;
                 case EventType.TaskScheduled:
                     historyEvent = new TaskScheduledEvent(eventId)
@@ -158,6 +191,7 @@ namespace DurableTask.SqlServer
                         Input = GetPayloadText(reader),
                         Name = GetName(reader),
                         Version = GetVersion(reader),
+                        ParentTraceContext = GetTraceContext(reader),
                     };
                     break;
                 case EventType.TimerCreated:
@@ -167,11 +201,17 @@ namespace DurableTask.SqlServer
                     };
                     break;
                 case EventType.TimerFired:
-                    historyEvent = new TimerFiredEvent(eventId)
+                    historyEvent = new TimerFiredEvent(eventId: -1)
                     {
                         FireAt = GetVisibleTime(reader),
                         TimerId = GetTaskId(reader),
                     };
+                    break;
+                case EventType.ExecutionSuspended:
+                    historyEvent = new ExecutionSuspendedEvent(eventId, GetPayloadText(reader));
+                    break;
+                case EventType.ExecutionResumed:
+                    historyEvent = new ExecutionResumedEvent(eventId, GetPayloadText(reader));
                     break;
                 default:
                     throw new InvalidOperationException($"Don't know how to interpret '{eventType}'.");
@@ -180,6 +220,20 @@ namespace DurableTask.SqlServer
             historyEvent.Timestamp = GetTimestamp(reader);
             historyEvent.IsPlayed = isOrchestrationHistory && (bool)reader["IsPlayed"];
             return historyEvent;
+        }
+
+        static bool TryGetFailureDetails(DbDataReader reader, out FailureDetails? details)
+        {
+            // Failure details are expected to be in JSON format. In previous versions, it was just an error message
+            // that isn't expected to be JSON.
+            string? text = GetPayloadText(reader);
+            if (string.IsNullOrEmpty(text) || text![0] != '{')
+            {
+                details = null;
+                return false;
+            }
+
+            return DTUtils.TryDeserializeFailureDetails(text, out details);
         }
 
         public static OrchestrationState GetOrchestrationState(this DbDataReader reader)
@@ -196,12 +250,13 @@ namespace DurableTask.SqlServer
                     }
                 };
             }
-            return new OrchestrationState
+
+            var state = new OrchestrationState
             {
-                CompletedTime = reader.GetUtcDateTimeOrNull(reader.GetOrdinal("CompletedTime")) ?? default,
-                CreatedTime = reader.GetUtcDateTimeOrNull(reader.GetOrdinal("CreatedTime")) ?? default,
+                CompletedTime = GetUtcDateTime(reader, "CompletedTime"),
+                CreatedTime = GetUtcDateTime(reader, "CreatedTime"),
                 Input = reader.GetStringOrNull(reader.GetOrdinal("InputText")),
-                LastUpdatedTime = reader.GetUtcDateTimeOrNull(reader.GetOrdinal("LastUpdatedTime")) ?? default,
+                LastUpdatedTime = GetUtcDateTime(reader, "LastUpdatedTime"),
                 Name = GetName(reader),
                 Version = GetVersion(reader),
                 OrchestrationInstance = new OrchestrationInstance
@@ -210,10 +265,27 @@ namespace DurableTask.SqlServer
                     ExecutionId = GetExecutionId(reader),
                 },
                 OrchestrationStatus = GetRuntimeStatus(reader),
-                Output = GetStringOrNull(reader, reader.GetOrdinal("OutputText")),
                 Status = GetStringOrNull(reader, reader.GetOrdinal("CustomStatusText")),
                 ParentInstance = parentInstance
             };
+
+            // The OutputText column is overloaded to contain either orchestration output or failure details
+            // if the task hub worker is configured to use legacy error propagation.
+            string? rawOutput = GetStringOrNull(reader, reader.GetOrdinal("OutputText"));
+            if (rawOutput != null)
+            {
+                if (state.OrchestrationStatus == OrchestrationStatus.Failed &&
+                    DTUtils.TryDeserializeFailureDetails(rawOutput, out FailureDetails? failureDetails))
+                {
+                    state.FailureDetails = failureDetails;
+                }
+                else
+                {
+                    state.Output = rawOutput;
+                }
+            }
+
+            return state;
         }
 
         internal static DateTime? GetVisibleTime(HistoryEvent historyEvent)
@@ -345,25 +417,77 @@ namespace DurableTask.SqlServer
 
         static DateTime GetVisibleTime(DbDataReader reader)
         {
-            int ordinal = reader.GetOrdinal("VisibleTime");
-            return GetUtcDateTime(reader, ordinal);
+            return GetUtcDateTime(reader, "VisibleTime");
         }
 
         static DateTime GetTimestamp(DbDataReader reader)
         {
-            int ordinal = reader.GetOrdinal("Timestamp");
-            return GetUtcDateTime(reader, ordinal);
+            return GetUtcDateTime(reader, "Timestamp");
         }
 
-        static DateTime? GetUtcDateTimeOrNull(this DbDataReader reader, int columnIndex)
+        static DateTime GetUtcDateTime(DbDataReader reader, string columnName)
         {
-            return reader.IsDBNull(columnIndex) ? (DateTime?)null : GetUtcDateTime(reader, columnIndex);
+            int ordinal = reader.GetOrdinal(columnName);
+            return GetUtcDateTime(reader, ordinal);
         }
 
         static DateTime GetUtcDateTime(DbDataReader reader, int ordinal)
         {
+            if (reader.IsDBNull(ordinal))
+            {
+                // Note that some serializers (like protobuf) won't accept non-UTC DateTime objects.
+                return DateTime.SpecifyKind(default, DateTimeKind.Utc);
+            }
+
             // The SQL client always assumes DateTimeKind.Unspecified. We need to modify the result so that it knows it is UTC.
             return DateTime.SpecifyKind(reader.GetDateTime(ordinal), DateTimeKind.Utc);
+        }
+
+        internal static SqlString GetTraceContext(HistoryEvent e)
+        {
+            if (e is not ISupportsDurableTraceContext eventWithTraceContext ||
+                eventWithTraceContext.ParentTraceContext == null)
+            {
+                return SqlString.Null;
+            }
+
+            DistributedTraceContext traceContext = eventWithTraceContext.ParentTraceContext;
+
+            // We prefer a simple format instead of JSON because external callers may interact with this
+            // data and we don't want to expose them to some internal JSON serialization format.
+            var sb = new StringBuilder(traceContext.TraceParent, capacity: 800);
+            if (!string.IsNullOrEmpty(traceContext.TraceState))
+            {
+                sb.Append('\n').Append(traceContext.TraceState);
+            }
+
+            return sb.ToString();
+        }
+
+        static DistributedTraceContext? GetTraceContext(DbDataReader reader)
+        {
+            int ordinal = reader.GetOrdinal("TraceContext");
+            if (reader.IsDBNull(ordinal))
+            {
+                return null;
+            }
+
+            string text = reader.GetString(ordinal);
+            if (string.IsNullOrEmpty(text))
+            {
+                return null;
+            }
+
+            string[] parts = text.Split(TraceContextSeparators, count: 2, StringSplitOptions.RemoveEmptyEntries);
+            var traceContext = new DistributedTraceContext(traceParent: parts[0]);
+
+            if (parts.Length > 1)
+            {
+                traceContext.TraceState = parts[1];
+            }
+
+            traceContext.ActivityStartTime = GetTimestamp(reader);
+            return traceContext;
         }
 
         public static SqlParameter AddInstanceIDsParameter(
@@ -436,7 +560,7 @@ namespace DurableTask.SqlServer
             var context = new SprocExecutionContext();
             try
             {
-                return await WithRetry(() => executor(command), context, traceHelper, instanceId);
+                return await WithRetry(command, executor, context, traceHelper, instanceId);
             }
             finally
             {
@@ -486,7 +610,7 @@ namespace DurableTask.SqlServer
             }
         }
 
-        static async Task<T> WithRetry<T>(Func<Task<T>> func, SprocExecutionContext context, LogHelper traceHelper, string? instanceId, int maxRetries = 5)
+        static async Task<T> WithRetry<T>(DbCommand command, Func<DbCommand, Task<T>> executor, SprocExecutionContext context, LogHelper traceHelper, string? instanceId, int maxRetries = 5)
         {
             context.RetryCount = 0;
 
@@ -494,7 +618,12 @@ namespace DurableTask.SqlServer
             {
                 try
                 {
-                    return await func();
+                    // Open connection if network blip caused it to close on a previous attempt
+                    if (command.Connection.State != ConnectionState.Open)
+                    {
+                        await command.Connection.OpenAsync();
+                    }
+                    return await executor(command);
                 }
                 catch (Exception e)
                 {